Fix sass-rails v6 stylesheets issue
sassc handles imports differently than it has in the past.

We previously had application.scss import Bootstrap, then import announcements and the announcements file depended on bootstrap defining variables which got passed down into the same scope that announcements was included in.

It seems this has changed, so announcements can't rely upon the import anymore and must explicitly import the dependencies.

The error Rails gives us is that the application.css isn't available for precompile, so it's silently failing on the variable reference. This appears to be a bug in sassc-ruby or something that is silently failing on the compile rather than raising an error about the missing variable.